Fetch and cache IPs lazily instead of upon initialization
By fetching and caching the IPs lazily, the app initialization does not depend on a cache store to be operational. A cache store cannot be expected to be always present upon app initialization (for example when running a rake task).

Taken from the open PR at https://github.com/modosc/cloudflare-rails/pull/52